[Enhance] refactoring 1
