+++ src/routes/admin.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const { isAdmin } = require("../middleware");
-const catchAsync = require("../utils/catchAsync");
+const catchAsync = require("../src/utils/catchAsync");
 
 const admin = require("../controllers/admin");
