+++ src/routes/users.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router({ mergeParams: true });
-const catchAsync = require("../utils/catchAsync");
+const catchAsync = require("../src/utils/catchAsync");
 const users = require("../controllers/users");
 const { isLoggedIn } = require("../middleware");
